requirements.txt
attrs==16.3.0
backoff==1.8.0
certifi==2020.6.20
chardet==3.0.4
ciso8601==2.1.3
idna==2.10
jsonschema==2.6.0
pendulum==1.2.0
python-dateutil==2.8.1
pytz==2018.4
pytzdata==2020.1
requests==2.24.0
simplejson==3.11.1
singer-python==5.9.0
singer-tools==0.4.1
six==1.15.0
strict-rfc3339==0.7
tenacity==6.2.0
terminaltables==3.1.0
tzlocal==2.1
urllib3==1.25.10
